# modules/core-platform/api/v1/openapi.yaml
...
openapi: 3.1.0
...
paths:
...
  /api/v1/checkout/validate-address:
    post:
      summary: Validate shipping address
      description: |
        Validates a shipping address using USPS or Lob API. Returns normalized address
        if valid, or validation errors if invalid.

        **Use Case:** Call this during checkout step 1 to verify address before proceeding.
        **Feature Flag:** Respects `checkout.address-validation.enabled` flag.
        **Authentication:** Optional - supports both guest and authenticated sessions.
      operationId: validateAddress
      tags:
      - Storefront
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: Address validation completed (may contain errors)
          content:
            application/json: {}
        "400":
          description: Invalid request format
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
...
info:
  title: Village Storefront API
  version: 1.0.0
  description: Multi-tenant SaaS ecommerce platform API
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value